move ValueProxy to it's own file
